feat: cambio de int a double para el escaneo del saldo del qr

+++ lib/models/transaccion_request.dart
@@ -2,7 +2,7 @@ class TransaccionRequest {
   final String traUuid;
   final String usuUuid;
   final String traMetodoPago;
-  final int traAmount;
+  final double traAmount;
   final String tokenUsuario;
 
   TransaccionRequest({
